recyclique-1.4.4/bot/requirements.txt
python-telegram-bot==20.7
redis==5.0.1
pydantic==2.5.0
pydantic-settings==2.1.0
httpx==0.25.2
fastapi==0.104.1
uvicorn==0.24.0
